requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
black==24.4.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.10.0
uvicorn==0.29.0
websockets==12.0
wsproto==1.2.0
